+++ CHANGELOG.md
@@ -4,6 +4,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.7.0]
+
+### Changed
+- Swift 4.2
+- animate thumb movement to make it smoother.
+
+### Fixed
+- zero-width sliding area in iOS 12 caused all thumbs to be shown on left and crash on drag.
+
+## [1.6.0] - 2018-07-06
+
+### Added
+- implement `isContinuous`.
+- send `primaryActionTriggered` event in addition to `valueChanged`.
+
+## [1.5.1] - 2018-07-03
+
+### Fixed
+- correctly handle non-zero `minimumValue` in horizontal orientation.
+
+## [1.5.0] - 2018-07-01
+
+### Added
+- send `.touchUpInside` event when user finished dragging. (thanks benjaminfischer!)
+
 ## [1.4.3] - 2018-06-25
 
 ### Fixed